Move ObjC and JS bridge testing into it's own file
